Added support for /interrogate endpoint and fixed some validation

+++ schemas/img2img.py
@@ -168,7 +168,7 @@ IMG2IMG_SCHEMA = {
         'default': ''
     },
     'refiner_switch_at': {
-        'type': int,
+        'type': float,
         'required': False,
         'default': 0
     },
